cleaned up and renamed format_brawler_data.py
