--- cronjobs_and_ml/output/formatted_brawler_data.csv
@@ -1,82 +1,0 @@
-brawler_id,win_rate,usage_rate,standardized_winrate,standardized_usage_rate,composite_score,rank,class
-Jacky,0.579,0.012,2.058,-0.034,2.057,1.0,tank
-Eve,0.572,0.002,1.912,-0.807,1.904,2.0,damage_dealer
-Nita,0.561,0.027,1.666,1.154,1.678,3.0,damage_dealer
-Meg,0.559,0.007,1.63,-0.384,1.626,4.0,tank
-Doug,0.557,0.003,1.58,-0.695,1.573,5.0,support
-Chuck,0.555,0.006,1.534,-0.473,1.53,6.0,damage_dealer
-Buster,0.554,0.002,1.517,-0.763,1.51,7.0,tank
-Frank,0.551,0.031,1.458,1.456,1.472,8.0,tank
-Jessie,0.55,0.048,1.417,2.746,1.445,9.0,controller
-Bibi,0.546,0.021,1.333,0.697,1.34,10.0,tank
-Sandy,0.541,0.019,1.232,0.527,1.237,11.0,controller
-Surge,0.535,0.018,1.094,0.449,1.098,12.0,damage_dealer
-Gale,0.53,0.013,0.986,0.019,0.986,13.0,controller
-Chester,0.527,0.007,0.911,-0.385,0.907,14.0,damage_dealer
-Berry,0.526,0.019,0.891,0.504,0.896,15.0,support
-Mr. P,0.524,0.003,0.842,-0.754,0.834,16.0,controller
-Crow,0.52,0.033,0.774,1.62,0.79,17.0,assassin
-Mico,0.518,0.006,0.731,-0.48,0.727,18.0,assassin
-Lily,0.515,0.009,0.662,-0.224,0.66,19.0,assassin
-Melodie,0.515,0.01,0.654,-0.144,0.652,20.0,assassin
-Colette,0.512,0.026,0.582,1.025,0.592,21.0,damage_dealer
-R-T,0.51,0.001,0.544,-0.838,0.536,22.0,damage_dealer
-Tick,0.509,0.015,0.525,0.172,0.526,23.0,thrower
-Edgar,0.506,0.057,0.464,3.463,0.498,24.0,assassin
-Bo,0.505,0.019,0.441,0.504,0.446,25.0,controller
-Tara,0.505,0.006,0.432,-0.468,0.427,26.0,damage_dealer
-Angelo,0.502,0.009,0.361,-0.282,0.358,27.0,sniper
-Fang,0.501,0.019,0.351,0.507,0.356,28.0,assassin
-Belle,0.5,0.007,0.334,-0.426,0.329,29.0,sniper
-Larry & Lawrie,0.497,0.009,0.265,-0.251,0.262,30.0,thrower
-Amber,0.496,0.007,0.23,-0.412,0.226,31.0,controller
-Cordelius,0.494,0.008,0.186,-0.323,0.183,32.0,assassin
-Draco,0.493,0.007,0.168,-0.442,0.163,35.0,tank
-Griff,0.493,0.01,0.165,-0.2,0.163,36.0,controller
-Buzz,0.493,0.027,0.161,1.168,0.173,33.0,assassin
-Leon,0.492,0.039,0.147,2.045,0.168,34.0,assassin
-Grom,0.491,0.004,0.132,-0.651,0.126,37.0,thrower
-Bull,0.491,0.007,0.118,-0.428,0.114,39.0,tank
-Rico,0.49,0.026,0.114,1.035,0.124,38.0,damage_dealer
-Kit,0.486,0.004,0.026,-0.645,0.02,40.0,support
-Emz,0.485,0.031,0.001,1.447,0.015,41.0,controller
-Byron,0.485,0.006,-0.007,-0.472,-0.012,43.0,support
-Mandy,0.484,0.006,-0.025,-0.488,-0.03,44.0,sniper
-Pearl,0.484,0.004,-0.031,-0.634,-0.037,45.0,damage_dealer
-Charlie,0.484,0.002,-0.032,-0.766,-0.039,46.0,controller
-Colt,0.484,0.05,-0.035,2.883,-0.006,42.0,damage_dealer
-Dynamike,0.481,0.045,-0.091,2.556,-0.065,47.0,thrower
-Brock,0.479,0.031,-0.134,1.421,-0.12,48.0,sniper
-El Primo,0.478,0.022,-0.164,0.743,-0.157,49.0,tank
-Lou,0.478,0.005,-0.167,-0.596,-0.173,50.0,controller
-Poco,0.475,0.007,-0.219,-0.381,-0.223,52.0,support
-Piper,0.475,0.021,-0.221,0.664,-0.214,51.0,sniper
-Maisie,0.474,0.003,-0.238,-0.715,-0.245,54.0,sniper
-Spike,0.474,0.04,-0.257,2.139,-0.235,53.0,damage_dealer
-Nani,0.471,0.007,-0.311,-0.411,-0.315,55.0,sniper
-Otis,0.471,0.001,-0.325,-0.855,-0.333,56.0,controller
-Rosa,0.468,0.004,-0.391,-0.667,-0.398,57.0,tank
-Shelly,0.466,0.016,-0.418,0.25,-0.415,58.0,damage_dealer
-Mortis,0.466,0.009,-0.432,-0.267,-0.435,59.0,assassin
-Carl,0.466,0.004,-0.433,-0.631,-0.44,60.0,damage_dealer
-Gray,0.452,0.002,-0.723,-0.779,-0.731,61.0,support
-Gene,0.45,0.006,-0.768,-0.511,-0.773,62.0,controller
-Janet,0.449,0.001,-0.8,-0.899,-0.809,63.0,sniper
-Sprout,0.447,0.002,-0.84,-0.778,-0.847,64.0,thrower
-Squeak,0.445,0.005,-0.893,-0.557,-0.898,65.0,controller
-Penny,0.443,0.008,-0.924,-0.367,-0.928,66.0,controller
-Barley,0.441,0.007,-0.972,-0.381,-0.976,67.0,thrower
-Max,0.441,0.01,-0.98,-0.16,-0.981,68.0,support
-Ash,0.44,0.001,-0.993,-0.866,-1.002,69.0,tank
-Bonnie,0.437,0.002,-1.055,-0.771,-1.063,70.0,sniper
-Sam,0.437,0.001,-1.06,-0.864,-1.068,71.0,assassin
-Bea,0.436,0.004,-1.091,-0.644,-1.097,72.0,sniper
-Ruffs,0.435,0.001,-1.111,-0.9,-1.12,73.0,support
-Lola,0.431,0.002,-1.189,-0.816,-1.198,74.0,damage_dealer
-Willow,0.431,0.002,-1.2,-0.832,-1.209,75.0,controller
-8-Bit,0.428,0.009,-1.272,-0.248,-1.274,76.0,damage_dealer
-Darryl,0.421,0.009,-1.413,-0.293,-1.416,77.0,tank
-Stu,0.418,0.004,-1.47,-0.619,-1.477,78.0,assassin
-Hank,0.398,0.001,-1.916,-0.87,-1.925,79.0,tank
-Gus,0.359,0.002,-2.785,-0.811,-2.793,80.0,support
-Pam,0.338,0.002,-3.239,-0.836,-3.247,81.0,support
